Initial
Knowledge Tests is an Quiz like app intended to be used to take tests
from users and the users give themselves scores after check the correct
answer.

This commit is the initial state of the project where I'm testing
Spring Security rules.

At this time, we have:
  - A working API;
  - Auth via Spring Security + Oauth2 (resource server)
and we need to:
  - Make JWT autority entries to be converted to somewhat Spring
    Security understands;
  - Everything else;